requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anonfile==0.2.7
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
eventlet==0.33.3
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==4.0.1
loguru==0.7.0
multidict==6.0.4
numpy==1.24.3
opencv-python==4.7.0.72
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-engineio==4.4.1
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.8.0
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tqdm==4.60.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchgod==0.8.2
websocket-client==1.5.1
websockets==11.0.2
yarl==1.9.2
zipp==3.15.0
